integrate lib/lineinput.rb to lib/review/lineinput.rb. Closes: #1600
